Updates on enhanced api (#194)
