+++ tests/flytekit/unit/use_scenarios/unit_testing/test_type_hints.py
@@ -1,26 +1,31 @@
+import inspect
 import typing
 
 import flytekit
-from flytekit import typing as _flytekit_typing
-from flytekit.annotated import stuff, context_manager
-from flytekit.sdk.tasks import python_task, inputs
-from flytekit.sdk.test_utils import flyte_test
+import flytekit.annotated.task
+import flytekit.annotated.workflow
+from flytekit.annotated import context_manager
+from flytekit.annotated.task import task
+from flytekit.annotated.workflow import workflow
+from flytekit.annotated.interface import extract_return_annotation, transform_variable_map
 
 
 def test_default_wf_params_works():
-    @stuff.task
+    @task
     def my_task(a: int):
         wf_params = flytekit.current_context()
         assert wf_params.execution_id == 'ex:local:local:local'
+
     my_task(a=3)
 
 
 def test_simple_input_output():
-    @stuff.task
+    @task
     def my_task(a: int) -> typing.NamedTuple("OutputsBC", b=int, c=str):
         ctx = flytekit.current_context()
         assert ctx.execution_id == 'ex:local:local:local'
-        return a+2, "hello world"
+        return a + 2, "hello world"
+
     assert my_task(a=3) == (5, 'hello world')
 
     # ctx = context_manager.FlyteContext.current_context()
@@ -31,7 +36,7 @@ def test_simple_input_output():
 
 
 def test_simple_input_no_output():
-    @stuff.task
+    @task
     def my_task(a: int):
         pass
 
@@ -44,10 +49,11 @@ def test_simple_input_no_output():
 
 
 def test_single_output():
-    @stuff.task
+    @task
     def my_task() -> str:
         return "Hello world"
-    assert my_task() =='Hello world'
+
+    assert my_task() == 'Hello world'
 
     ctx = context_manager.FlyteContext.current_context()
     with ctx.new_compilation_context() as ctx:
@@ -60,17 +66,18 @@ def test_single_output():
 
 def test_named_tuples():
     nt1 = typing.NamedTuple("NT1", x_str=str, y_int=int)
+
     def x(a: int, b: str) -> typing.NamedTuple("NT1", x_str=str, y_int=int):
         return ("hello world", 5)
 
     def y(a: int, b: str) -> nt1:
         return nt1("hello world", 5)
 
-    result = stuff.get_output_variable_map(x.__annotations__)
+    result = transform_variable_map(extract_return_annotation(inspect.signature(x).return_annotation))
     assert result['x_str'].type.simple == 3
     assert result['y_int'].type.simple == 1
 
-    result = stuff.get_output_variable_map(y.__annotations__)
+    result = transform_variable_map(extract_return_annotation(inspect.signature(y).return_annotation))
     assert result['x_str'].type.simple == 3
     assert result['y_int'].type.simple == 1
 
@@ -79,7 +86,7 @@ def test_unnamed_typing_tuple():
     def z(a: int, b: str) -> typing.Tuple[int, str]:
         return 5, "hello world"
 
-    result = stuff.get_output_variable_map(z.__annotations__)
+    result = transform_variable_map(extract_return_annotation(inspect.signature(z).return_annotation))
     assert result['out_0'].type.simple == 1
     assert result['out_1'].type.simple == 3
 
@@ -88,7 +95,7 @@ def test_regular_tuple():
     def q(a: int, b: str) -> (int, str):
         return 5, "hello world"
 
-    result = stuff.get_output_variable_map(q.__annotations__)
+    result = transform_variable_map(extract_return_annotation(inspect.signature(q).return_annotation))
     assert result['out_0'].type.simple == 1
     assert result['out_1'].type.simple == 3
 
@@ -97,20 +104,20 @@ def test_single_output_new_decorator():
     def q(a: int, b: str) -> int:
         return a + len(b)
 
-    result = stuff.get_output_variable_map(q.__annotations__)
+    result = transform_variable_map(extract_return_annotation(inspect.signature(q).return_annotation))
     assert result['out_0'].type.simple == 1
 
 
 def test_wf1():
-    @stuff.task
+    @task
     def t1(a: int) -> typing.NamedTuple("OutputsBC", t1_int_output=int, c=str):
-        return a+2, "world"
+        return a + 2, "world"
 
-    @stuff.task
+    @task
     def t2(a: str, b: str) -> str:
         return b + a
 
-    @stuff.workflow
+    @workflow
     def my_wf(a: int, b: str) -> (int, str):
         x, y = t1(a=a)
         d = t2(a=y, b=b)
@@ -125,15 +132,15 @@ def test_wf1():
 
 
 def test_wf1_run():
-    @stuff.task
+    @task
     def t1(a: int) -> typing.NamedTuple("OutputsBC", t1_int_output=int, c=str):
         return a + 2, "world"
 
-    @stuff.task
+    @task
     def t2(a: str, b: str) -> str:
         return b + a
 
-    @stuff.workflow
+    @workflow
     def my_wf(a: int, b: str) -> (int, str):
         x, y = t1(a=a)
         d = t2(a=y, b=b)
@@ -144,7 +151,6 @@ def test_wf1_run():
         'out_0': 7,
         'out_1': "hello world",
     }
-
 
 # def test_normal_path():
 #     # Write some random numbers to a file
@@ -162,8 +168,6 @@ def test_wf1_run():
 #
 
 
-
-
 # @flyte_test
 # def test_single_output():
 #     @python_task
@@ -172,7 +176,7 @@ def test_wf1_run():
 #             fh.writelines("hello world")
 #             my_output = upload_to_location(fh, "s3://my-known-location", format="csv")
 #             return fh
-    
+
 #     assert my_task.unit_test() == {'output': 'Hello world'}
 
 
@@ -184,7 +188,7 @@ def test_wf1_run():
 #     def my_task(ctx, fh: typing.BinaryIO):
 #         lines = fh.readlines()
 #         # assert 
-    
+
 #     # Option 1.1
 #     # To call the task for unit testing, users need to open a file and pass the handle
 #     with open('/mytest', mode='rb') as fh:
@@ -194,7 +198,7 @@ def test_wf1_run():
 #     # Option 1.2
 #     # Users pass a Path-Like object that flyte knows how to interpret and open or youo
 #     assert my_task.unit_test(fh=CustomPathLike('/mytest', format="csv")) == {}
-    
+
 #     # Option 2
 #     # Users receive a Path-Like type as a parameter to their function (much like how Blobs work today)
 #     # Option 2.1
